// src/enum/sys/comp-broadcast-types.enum.ts
export enum CompBroadCastTypes {
    Login = 0,
    Logout = 1,
    ModalClosed = 2,
    //ValidationComplete = 3,
    NeedRefresh = 4,
    ItemSelected = 5,
    Open = 6,
    ItemAdded = 7,
    Scroll = 8,
    Saved = 9,
    SaveAttempt = 10,
    Deleted = 11,
    GoogleAnalytics = 12,
    Busy = 13,
    Ready = 14
}